test(toggle): migrate enable-on-off-labels
